Updated Chapter 1 files

+++ CHAPTER 1/05_inputfromuser.py
@@ -1,5 +1,5 @@
 a = int( input("enter the 1 no:"))
-b = int( input("enter the 2 no:"))
+b = float( input("enter the 2 no:"))
 
 
 print("1 no is",a)
